Move casenotification to notify entities
