+++ src/main/java/uk/gov/hmcts/reform/enforcement/notify/repository/NotificationRepository.java
@@ -1,7 +1,7 @@
 package uk.gov.hmcts.reform.enforcement.notify.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import uk.gov.hmcts.reform.enforcement.notify.domain.CaseNotification;
+import uk.gov.hmcts.reform.enforcement.notify.entities.CaseNotification;
 
 import java.util.Optional;
 import java.util.UUID;
